Changing Role @Table name.

+++ src/main/java/br/com/ffroliva/portfolio/model/User.java
@@ -1,27 +1,17 @@
 package br.com.ffroliva.portfolio.model;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import javax.persistence.UniqueConstraint;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotNull;
-
+import br.com.ffroliva.portfolio.model.audit.DateAudit;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import br.com.ffroliva.portfolio.model.audit.DateAudit;
 
-import com.fasterxml.jackson.annotation.JsonManagedReference;
+import javax.persistence.*;
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotNull;
+import java.util.ArrayList;
+import java.util.List;
 
 @ToString
 @Entity
@@ -36,14 +26,14 @@ import com.fasterxml.jackson.annotation.JsonManagedReference;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor(staticName = "of")
-public class User extends DateAudit<Long> {
+public class User extends DateAudit<Integer> {
 
 	private static final long serialVersionUID = -4373524336039096881L;
 
 	@Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "user_id", updatable = false, nullable = false)
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Column(name="username", nullable = false)
