+++ src/main/java/br/com/ffroliva/portfolio/repository/UserRepository.java
@@ -1,16 +1,15 @@
 package br.com.ffroliva.portfolio.repository;
 
-import java.util.List;
-import java.util.Optional;
-
+import br.com.ffroliva.portfolio.model.User;
+import br.com.ffroliva.portfolio.model.UserRole;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
-import br.com.ffroliva.portfolio.model.User;
-import br.com.ffroliva.portfolio.model.UserRole;
+import java.util.List;
+import java.util.Optional;
 
-public interface UserRepository extends PagingAndSortingRepository<User, Long> {
+public interface UserRepository extends PagingAndSortingRepository<User, Integer> {
 
     Optional<User> findByUsername(String username);
